[libgweather] Phase-of-moon support, calculation improvements
The icon name returned by weather_info_get_icon_name() includes
an indication of the current phase of the moon for nighttime
observations 'clear' and 'few-clouds'.  Icons that correspond
to those shipped in gnome-icon-theme are included as well.

The changes made for sunrise and sunset calculations should give
more meaningful results for those in or near the polar regions.

Closes: bgo#564598, bgo#569298