Add serialization capabilities to GWeatherLocation
Applications using libgweather want to store locations in GSettings
or data files, and then be able to retrieve them, in a way that minimizes
the difference between acquiring the location from a GWeatherLocationEntry
or from storage.
To do so, a new location level is introduced, GWEATHER_LOCATION_DETACHED,
which is used as a fallback in case we cannot recover the right location
from the database.