GWeatherLocation: use a global shared world location
In order to have GWeatherLocationEntry and GWeatherTimezoneMenu
usable by GtkBuilder, we need to avoid passing a GWeatherLocation
for world. To do that, we can use a shared world, that is just
referenced by gweather_location_new_world(). Sharing the world
automatically shares all descendent locations, so GWeatherLocation:location
continues to work.

Unfortunately, this means that the @use_regions parameter becomes
deprecated, and it is now always assumed TRUE.