API break: replace gweather_location_new_world() with gweather_location_get_world()
The only parameter to it was deprecated, and the name was misleading,
as it was not creating a new world location but always returning
the same one. Also, the memory management was wrong in case the
application freed the world location immediately.

At the same time, let's remove some old cruft in GWeatherLocation.