don't use non-standard 'd' suffix on float literal
C does not define the 'd' suffix for floating point literals.  In fact,
'double' is the default type of literals if not modified by 'f' (float)
or 'l' (long float).

Additionally: 'float' arguments are automatically upcasted to 'double'
when passing through a varargs list, so this wouldn't be needed in any
case.

This problem was causing failures to build with clang.

https://bugzilla.gnome.org/show_bug.cgi?id=720207