Adding detached location support and expanding range of cities
Add detached location support and remov segmentation fault. Use
geocoding to get the geocode places for user's invalid input and then
using coordinates of the selected place find the nearest weather station
and using that create a detached location and set it as internal
location. Better handling of "Loading.." string.

https://bugzilla.gnome.org/show_bug.cgi?id=731466