GWeatherLocationEntry: respect country information when converting from geocode
Turns out that geocode-glib does give us a valid country code, so
we should make use of it and avoid blunders like putting cities in
the wrong country.

It's still possible we put cities in the wrong timezone, but given
the weirdness of timezones inside countries (I'm looking at you, Arizona),
the solution is just more cities in the static db.

https://bugzilla.gnome.org/show_bug.cgi?id=753332