Remove LIYW weather station
The Aviano US Air Force Air Base no longer reports METAR data, and
Wikipedia does not list LIYW as a valid ICAO code.
Quite likely, it was merged with LIPA (the Aviano Airport, managed
by the NATO/Italian Air Force). The two stations are in the same
place anyway, so it's not a loss to remove LIYW.