build: Improve chances of using libgweather as a subproject
Do not use add_global_arguments(), and do not reference source and build
root directories.

Instead of the former, we can use add_project_arguments(); in place of
the latter, we should use the include_directories() object.

We also want to make the dependency on libgweather include the
introspection data, as it will be useful for projects including
libgweather that are not written in C.