Remove Yahoo! Weather support
The RSS feed URL was moved from:
https://weather.yahooapis.com/forecastrss
to:
https://weather-ydn-yql.media.yahoo.com/forecastrss
as per https://developer.yahoo.com/weather/documentation.html

But this new location requires use of OAuth to link the API to the
application. Given that we relied on the URL to be freely accessible,
and that we have so few locations with a Yahoo "Where On Earth" ID set,
we might as well remove the code to avoid failing to fetch the weather
for those locations.

https://bugzilla.gnome.org/show_bug.cgi?id=791084