locations: Fix city attached to Chernivtsi airport
Remove Chagor (Chahor) from the list of cities, as it's a village:
https://en.wikipedia.org/wiki/Chahor

And attach the town of Chernivtsi to its airport instead, as that's
where it is anyway:
https://en.wikipedia.org/wiki/Chernivtsi
https://en.wikipedia.org/wiki/Chernivtsi_International_Airport

This also changes the spelling of the city and airport from Chernovsty
and makes them translatable, as they have slightly different
transliterations in other languages (eg. Czernowitz in German, Cernăuți
in Romanian).

https://bugzilla.gnome.org/show_bug.cgi?id=620897