GWeatherLocation: Remove _get_world_for_path() API
And the changes it required from commit ad604b79. There are a number
of internal uses of gweather_location_get_world(), including in the
GWeatherTimezone code, and adding new API for the sake of our tests
feels wrong, and would complicate the public API for no reason.

Instead, allow using the LIBGWEATHER_LOCATIONS_PATH envvar to override
the location of the Locations.xml file. This makes it easy to use the
same API in our tests and installed.

https://bugzilla.gnome.org/show_bug.cgi?id=791317