locations: Remove Edmonton Municipal (CYXD) station
Edmonton Municipal CYXD weather station is not in an operation anymore.
Edmonton Blatchford, CXEC, is in roughly the same location as an old
CYXD but this station does not provide a weather condition information
and only temperature data so it is not a suitable replacement.

Use CYED and CYEG instead.

https://bugzilla.gnome.org/show_bug.cgi?id=770497