location-entry: Fix completion entries that can't be differentiated
In commit 6787563, we started stripping the country from geocode-glib
results. But the last element was not always the country, but other
differentiating information (such as the state, or province) if all the
results were in the same country. And the statement:
"remove the country part, which will be readded by libgweather anyway."
doesn't seem to be correct anymore.

This fixes the unusable completion popup when there's no differentiating
information between the results.

https://bugzilla.gnome.org/show_bug.cgi?id=791403