locations: Add Gijón, Spain
It's a large enough city with an airport close to it (Asturias Airport).
The capital city of that principality, Oviedo, is slightly smaller, and
the airport is further. The location entry will automatically lookup
Ovideo and attach it to the closest airport, so there should be no need
to add it manually.

https://bugzilla.gnome.org/show_bug.cgi?id=597989