GWeatherLocation: Fix refcount of deserialized named timezones
When deserialising named timezones, make sure to hand out a reference to
the GWeatherLocation, as all the callers (and the public API) expect a
reference to be handed out.

This fixes a crash in gnome-shell when garbage collection is triggered
and the returned location is unreferenced.

See https://bugzilla.redhat.com/show_bug.cgi?id=1577561