GWeatherLocation: Fix double-free with "nearest stations"
The nearest station code would try to (badly) copy the nearest station's
location to add it as a child, but the copy was shallow, and the string
pointers, and other allocated memory wasn't duplicated, leading to
double-frees and invalid memory accesses if the world was unref'ed.

This also fixes the newly created GWeatherLocation to use g_slice_* to
allocate the memory chunk, otherwise we'll have memory corruption when
we try to free it with g_slice_free() in the unref() code.