GWeatherLocation: Additional fixes for deserializing a city
Since weather stations are no longer attached to cities, deserializing
cities in format two leads to a detached location pointing to the weather
station and not the city. Try first to find the city by name and coordinates.

This also fixes gnome-shell's world clocks menu to not miss configured
cities from countries with many timezones