yrno: Fix forecasts never being available because invalid
yrno was never setting the "valid" flag on the main info struct, just on
each one of the info structs created to hold the forecasts, so it was
impossible to get the forecasts unless *another* provider succeeded.

This fixes fetching forecasts for Helsinki, as METAR fetching fails and
YR.NO didn't set the valid flag.

Closes: #11