location-entry: Sort locations in drop-down
The order of locations as returned by GWeatherLocation used to be sorted
by the localised name. However, that requires quite a lot of resources
and is not needed in many usecases.

As such, sort the resulting tree model by location name. Note that this
actually changes the behaviour slightly as Countries were grouped by
region before and are now all sorted by name.

To do so, add a new column that allows direct comparison for sorting
purposes. Note that this cannot be the same as the compare name, because
order needs to be reversed for sorting but must remain the same for
matching.