location-entry: Sort locations in drop-down
The order of locations as returned by GWeatherLocation used to be sorted
by the localised name. However, that requires quite a lot of resources
and is not needed in many usecases.

To ensure that the parent takes precedence, it is now changed to be at
the front for sorting purposes.

As such, sort the resulting tree model by location name. Note that this
actually changes the behaviour slightly as Countries were grouped by
regsion before and are now all sorted by name.