Use explicit symbol annotation and visibility
All symbols should be hidden by default, and public symbols should be
explicitly annotated.

Once we commit to a stable API/ABI, we will be able to use versioned
macros to mark symbols for deprecation, and get compiler warnings that
can be toggled on and off by consumers of our API.

The current symbol visibility map includes two private symbols used by
our tests; to avoid this leakage, we can build libgweather as a static
library, and then use the static library for our tests and as the base
of our public shared library.