ci: Add (fallible) static analysis and asan jobs
Libgweather is a C library, and it's reading and parsing files, and
poking at web services: we should really be doing some minimum safety
checks that our code isn't just an exploit waiting to happen.

For the time being, both jobs are marked to allow failure, with the
explicit intent to lead us towards fixing the issues they find.