Reformat the code using clang-format
While it would be very nice to keep our artisanal, locally sourced,
hand-crafted coding style, our time is better spent trying to fix issues
and implement features, instead of doing coding style reviews. In order
to have our CI pipeline in charge of keeping and checking the coding
style of this project, we need to have a consistent baseline.

This consistency pass also gets rid of all the tabs, so we have that
going for us.