Avoid a potential NULL dereference
When asynchronously querying for the nearest location at the given
location we should keep a reference on the parent location, in case it
goes away mid-search.

This is especially important when passing `NULL` as the parent location:
the current code uses a `g_autoptr()` to store the "world" node, and
then passes the pointer unmodified; if that happened, the location would
be released at the end of the function, and the asynchronously invoked
callback would get a garbage pointer.