metar test: Don't use "known" duplicates
The metar list provided by NOAA contains duplicate ICAO codes.
Those duplicates may change over time, so it does not make sense
to use an array of known duplicates. Presently a duplicate is rejected
when it is known, and inserted if it is not, leading to  a difference
in the number of insertions and the number of keys, so that the test
fails. If any duplicate is rejected, the number of insertions is
always equal to the number of keys (unless something goes wrong
in g_hash_table_insert). And this is the default behavior for known
duplicates anyway.

Closes: https://gitlab.gnome.org/GNOME/libgweather/-/issues/168