Fix fallback metric unit detection logic
When HAVE__NL_MEASUREMENT_MEASUREMENT is not defined (for example on
musl systems), the fallback logic checks for units in translation files.

If the unit in the translation file is metric, we should use metric and
not print a warning about wrong translation and then use metric.

Introduced in 1c140fc8ce08260d5008847945bf345654ad7fa8