search: add simple search engine for GWeatherLocation
This is a trigrams-based approach to searching across a normalized and
casefolded corpus.

The first pass you find all possible documents which contain all of the
trigrams that have been searched for. After that, you filter out the items
doing a more traditional match. That sample set is generally a small
fraction of the entire corpus.

A static gweather-search-indexer is also compiled which can generate a
search index using libgweather to extract the current Locations.bin data.

It is probably ideal to use this at build time to generate the search
index and install it alongside the Locations.bin.