db: index city lat/lon using a kdtree
This creates a 3-dimentional kd-tree with latitude/longitude positions for
each of the cities. The goal here is to make find_nearest_city() quite fast
when deserializing a child of World.

We leave the other find_nearest_city() alone because the way they are
structured implies a subset query which can't be made fast without a
partitioning index.