Plug SoupMessage leak
Whenever we queue a SoupMessage for an HTTP query, we put it in a list
and acquire a reference to it. The reference is held until the request
ends, or it's cancelled.

The METAR and MetNo providers create a new SoupMessage instance, queue
it, but do not transfer the ownership to the GweatherInfo queue, which
means we end up leaking the SoupMessage and all its ancillary data,
including the URI object.